+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Import link.xml caused an infinite loop ([#315](https://github.com/getsentry/sentry-unity/pull/315))
 - Removed unused .asmdefs which clears a warning from console ([#316](https://github.com/getsentry/sentry-unity/pull/316))
+- Don't send negative line number ([#317](https://github.com/getsentry/sentry-unity/pull/317))
 
 ## 0.5.2